requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
hyperliquid-python-sdk>=0.4.0
token_data>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=7.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
